backend/requirements.txt
# ── Core ──
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Compress==1.23
pymongo==4.16.0
python-dotenv==1.2.1
requests==2.32.4
gunicorn==25.1.0
certifi==2025.8.3

# ── Scraping / PDF ──
beautifulsoup4==4.13.4
pdfplumber==0.11.9

# ── AI / ML ──
google-generativeai==0.8.6
scikit-learn==1.8.0
numpy==1.26.2
pandas==2.3.1
prophet==1.3.0

# ── Image Analysis ──
Pillow==12.1.1

# ── New features (Alerts, Calendar, Soil, Crop Recommender) ──
# No additional packages needed — uses existing Gemini + Open-Meteo APIs

# ── Yield model persistence ──
joblib==1.4.2
